+++ changelog.md
@@ -5,6 +5,7 @@
 ### Version 5.0.11
 
 * `cordapp`: Raise minimum Gradle version to 5.1.
+* `cordapp`: Refresh the certificate for the development CorDapp signing key.
 
 ### Version 5.0.10